pyproject.toml
[project]
name = "tablefree"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "keyring>=25.7.0",
    "mysql-connector-python>=9.6.0",
    "psycopg2-binary>=2.9.11",
    "pyside6>=6.10.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]
